refactor: Create ConfigApiClient with fetchConfig function
