+++ src/commonMain/kotlin/com/rakuten/tech/mobile/remoteconfig/ConfigFetcher.kt
@@ -1,13 +1,16 @@
 package com.rakuten.tech.mobile.remoteconfig
 
 import io.ktor.client.HttpClient
+import io.ktor.client.features.defaultRequest
 import io.ktor.client.features.json.JsonFeature
 import io.ktor.client.features.json.serializer.KotlinxSerializer
+import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.response.readText
 import io.ktor.client.response.HttpResponse
 import io.ktor.http.Url
+import io.ktor.http.isSuccess
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
@@ -15,50 +18,18 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlin.jvm.Transient
 
-internal expect val ApplicationDispatcher: CoroutineDispatcher
 
 class ConfigFetcher internal constructor(
-    baseUrl: String,
-    appId: String,
-    private val subscriptionKey: String,
-    private val scope: CoroutineScope
+    private val client: HttpClient,
+    baseUrl: String
 ) {
+    private val address = Url("${baseUrl}/config")
 
-    constructor(
-        baseUrl: String,
-        appId: String,
-        subscriptionKey: String
-    ) : this (
-        baseUrl = baseUrl,
-        appId = appId,
-        subscriptionKey = subscriptionKey,
-        scope = CoroutineScope(ApplicationDispatcher)
-    )
+    suspend fun fetch(): Config {
+        val response = client.get<HttpResponse>(address)
 
-    private val client = HttpClient {
-        install(JsonFeature) {
-            serializer = KotlinxSerializer()
-        }
-    }
-    private val address = Url("${baseUrl}/app/$appId/config")
-
-    fun fetch(response: (Config) -> Unit, error: (Exception) -> Unit) {
-        scope.launch {
-            try {
-                val config = fetchConfig()
-                response(config)
-            } catch (exception: ResponseException) {
-                error(exception)
-            }
-        }
-    }
-
-    private suspend fun fetchConfig(): Config {
-        val response = client.get<HttpResponse>(address) {
-            header("apiKey", "ras-$subscriptionKey")
-        }
-
-        if (response.status.value >= HTTP_STATUS_300) {
+        response.status.isSuccess()
+        if (!response.status.isSuccess()) {
             throw ResponseException(response)
         }
 
@@ -68,11 +39,6 @@ class ConfigFetcher internal constructor(
         val signature = response.headers["Signature"] ?: ""
 
         return Config(body, signature, keyId)
-    }
-
-    companion object {
-        private const val CACHE_SIZE = 1024 * 1024 * 2L
-        private const val HTTP_STATUS_300 = 300
     }
 }
 
@@ -100,11 +66,3 @@ data class Config(
         fun fromJsonString(body: String) = Json.nonstrict.parse(serializer(), body)
     }
 }
-
-/**
- * Base for default response exceptions.
- * @param response: origin response
- */
-open class ResponseException(
-    @Transient val response: HttpResponse
-) : IllegalStateException("Bad response: ${response.status}")
